+++ examples/npm-typescript/wirespec/user.ws
@@ -11,4 +11,3 @@ endpoint GetUsers GET /api/users -> {
     200 -> User[]
     500 -> UserError
 }
-
