+++ src/compiler/core/src/commonMain/kotlin/community/flock/wirespec/compiler/core/tokenize/TokenTypes.kt
@@ -21,7 +21,6 @@ data object Underscore : TokenType
 data object Character : TokenType
 data object Arrow : TokenType
 data object Pipe : TokenType
-data object At : TokenType
 data object LiteralString : TokenType
 data object EndOfProgram : TokenType {
     const val VALUE = "EOP"
