+++ src/compiler/core/src/commonTest/kotlin/community/flock/wirespec/compiler/core/parse/ParseAnnotationTest.kt
@@ -1,6 +1,7 @@
 package community.flock.wirespec.compiler.core.parse
 
 import arrow.core.nonEmptyListOf
+import community.flock.wirespec.compiler.core.FileUri
 import community.flock.wirespec.compiler.core.ModuleContent
 import community.flock.wirespec.compiler.core.ParseContext
 import community.flock.wirespec.compiler.core.WirespecSpec
@@ -17,7 +18,7 @@ class ParseAnnotationTest {
 
     private fun parser(source: String) = object : ParseContext, NoLogger {
         override val spec = WirespecSpec
-    }.parse(nonEmptyListOf(ModuleContent("", source))).map { it.modules.flatMap(Module::statements) }
+    }.parse(nonEmptyListOf(ModuleContent(FileUri(""), source))).map { it.modules.flatMap(Module::statements) }
 
     @Test
     fun testSimpleAnnotationOnType() {
@@ -259,7 +260,7 @@ class ParseAnnotationTest {
         """.trimMargin()
 
         parser(source)
-            .shouldBeRight { it.joinToString { it.message } }
+            .shouldBeRight { it -> it.joinToString { it.message } }
             .shouldHaveSize(1)
             .first()
             .shouldBeInstanceOf<Endpoint>()
