+++ src/compiler/core/src/commonTest/kotlin/community/flock/wirespec/compiler/core/tokenize/TestUtils.kt
@@ -1,5 +1,6 @@
 package community.flock.wirespec.compiler.core.tokenize
 
+import arrow.core.NonEmptyList
 import community.flock.wirespec.compiler.core.WirespecSpec
 import io.kotest.matchers.collections.shouldNotBeEmpty
 import io.kotest.matchers.shouldBe
@@ -17,19 +18,15 @@ fun testTokenizer(
         .let { it shouldBe expected }
 }
 
-fun Iterable<Token>.shouldBeSound(source: String): Iterable<Token> {
-    with(filter { it.type != EndOfProgram }) {
-        windowed(2, 1, false) { (a, b) ->
+fun NonEmptyList<Token>.shouldBeSound(source: String) = apply {
+    dropLast(1)
+        .also { tokens -> source.length shouldBe tokens.sumOf { it.coordinates.idxAndLength.length } }
+        .windowed(2) { (a, b) ->
             a.coordinates.idxAndLength.idx + b.coordinates.idxAndLength.length shouldBe b.coordinates.idxAndLength.idx
-            a.getEnd() shouldBe b.getStart()
+            a.coordinates.idxAndLength.idx shouldBe b.coordinates.getStartPos()
             a.value.length shouldBe a.coordinates.idxAndLength.length
             b.value.length shouldBe b.coordinates.idxAndLength.length
         }
-        source.length shouldBe sumOf { it.coordinates.idxAndLength.length }
-    }
-    return this
 }
 
 private fun Token.Coordinates.getStartPos() = idxAndLength.idx - idxAndLength.length
-private fun Token.getStart() = coordinates.getStartPos()
-private fun Token.getEnd() = coordinates.idxAndLength.idx
