+++ src/ide/intellij-plugin/src/main/kotlin/community/flock/wirespec/ide/intellij/Types.kt
@@ -47,6 +47,5 @@ interface Types {
 
         val ANNOTATION = ElementType("ANNOTATION")
         val LITERAL_STRING = ElementType("STRING_LITERAL")
-        val AT = ElementType("AT")
     }
 }
